+++ tests/functional/testdata/test_param/search_param.py
@@ -9,12 +9,12 @@ from functional.testdata.etl_indexes.persons_indexes import (
 from tests.functional.testdata.es_generate.film_generate import (
     generate_films
 )
-from tests.functional.testdata.es_generate.genre_generate import (
+from tests.functional.testdata.es_generate.person_generate import (
     generate_persons
 )
 
 FILMS_PARAM = [
-    'films',
+    'movies',
     'films/search/',
     generate_films(),
     MOVIES_INDEX_MAPPING
